replace visibleEntries with visibleSimpleEntriesForCompactList in editor
The purpose of visibleSimpleEntriesForCompactList is to try to improve AngularJS performance by reducing the number of watches.  MS Edge browser seems to be particularly suseptable to the inefficiencies of watches.  visibleSimpleEntriesForCompact list is essentially a list of custom objects containing just the data needed for displaying the list, not the entire list of entries.